+++ app/src/main/java/com/issever/issevercore/data/service/FactService.kt
@@ -3,11 +3,14 @@ package com.issever.issevercore.data.service
 import com.issever.issevercore.data.model.CatFactsResponse
 import retrofit2.Response
 import retrofit2.http.GET
+import retrofit2.http.Query
 
 interface FactService {
 
     @GET("facts")
-    suspend fun getFacts(): Response<CatFactsResponse>
+    suspend fun getFacts(
+        @Query("page") page: Int,
+    ): Response<CatFactsResponse>
 
     // This endpoint will return an error
     @GET("sample")