+++ isseverCore/src/main/java/com/issever/core/data/model/SnackbarMessage.kt
@@ -1,11 +1,11 @@
 package com.issever.core.data.model
 
-import com.issever.core.data.enums.SnackbarType
+import com.issever.core.data.enums.StateType
 
 
 data class SnackbarMessage(
     val message: String?= "",
-    val type: SnackbarType? = null,
+    val type: StateType? = null,
     val actionText: String? = "",
     val action: (() -> Unit)? = null
 )